Pythonic refactor: date_tool.py -> datetool.py
One of several refactors to remove the use of utility classes which only
contain static methods.
