+++ pd_utils/report/coverage_gap_report.py
@@ -8,7 +8,7 @@ from typing import Any
 
 from pd_utils.model import EscalationRuleCoverage as EscCoverage
 from pd_utils.model import ScheduleCoverage as SchCoverage
-from pd_utils.util import DateTool
+from pd_utils.util import datetool
 from pd_utils.util import IOUtil
 from pd_utils.util import PagerDutyAPI
 
@@ -33,8 +33,8 @@ class CoverageGapReport:
             max_query_limit: Number of objects to request at once from PD (max: 100)
             look_ahead_days: Number of days to look ahead on schedule (default: 14)
         """
-        self._since = DateTool.utcnow_isotime()
-        self._until = DateTool.add_offset(self._since, days=look_ahead_days)
+        self._since = datetool.utcnow_isotime()
+        self._until = datetool.add_offset(self._since, days=look_ahead_days)
         self._max_query_limit = max_query_limit
         self._schedule_map: dict[str, SchCoverage] = {}
         self._escalation_map: dict[str, EscCoverage] = {}
@@ -120,7 +120,7 @@ class CoverageGapReport:
 
         for ep_rule in self._escalation_map.values():
             times = self._extract_entries(ep_rule.rule_target_ids)
-            ep_rule.is_fully_covered = DateTool.is_covered(
+            ep_rule.is_fully_covered = datetool.is_covered(
                 time_slots=times,
                 range_start=self._since,
                 range_stop=self._until,
