+++ pd_utils/util/__init__.py
@@ -1,12 +1,10 @@
 from __future__ import annotations
 
-from .date_tool import DateTool
 from .io_util import IOUtil
 from .pagerduty_api import PagerDutyAPI
 from .runtime_init import RuntimeInit
 
 __all__ = [
-    "DateTool",
     "IOUtil",
     "PagerDutyAPI",
     "RuntimeInit",
